+++ src/Components/Attendees.js
@@ -1,6 +1,25 @@
 import Attendee from "./Attendee.js"
-export default function Attendees({attendee,event,updateEventAttendance}) {
+import { useState } from "react";
+export default function Attendees({event,updateEventAttendance,attendees}) {
+  const [showAttendees, setShowAttendees] = useState(false);
+  function toggleEventAttendees() {
+    setShowAttendees(!showAttendees);
+  }
   return (
-    <Attendee attendee={attendee} event={event} updateEventAttendance={updateEventAttendance}/>
+    
+    <>
+      <button onClick={toggleEventAttendees}>
+        {!showAttendees ? "Show Attendees" : "Hide Attendees"}
+      </button>
+
+      {showAttendees ? (
+        <div className="attendees">
+          {attendees.map((attendee, index) => (
+            <Attendee event={event} toggleEventAttendees={toggleEventAttendees} showAttendees={showAttendees} attendee={attendee} updateEventAttendance={updateEventAttendance} />
+          ))}
+        </div>
+      ) : null}
+    </>
+    // <Attendee attendee={attendee} event={event} updateEventAttendance={updateEventAttendance}/>
   )
 }
