+++ src/Components/Event.js
@@ -1,5 +1,5 @@
 import Attendees from "./Attendees.js"
-export default function Event({event,toggleEventAttendees,showAttendees,attendees,updateEventAttendance}) {
+export default function Event({event,updateEventAttendance,toggleEventAttendees,showAttendees,attendees}) {
   return (
   <>
     <li key={event.id}>
@@ -10,19 +10,7 @@ export default function Event({event,toggleEventAttendees,showAttendees,attendee
       <br />
       <span>Organized by: {event.organizer} </span>
       <br />
-      <>
-        <button onClick={toggleEventAttendees}>
-          {!showAttendees ? "Show Attendees" : "Hide Attendees"}
-        </button>
-
-        {showAttendees ? (
-          <div className="attendees">
-            {attendees.map((attendee, index) => (
-              <Attendees event={event} toggleEventAttendees={toggleEventAttendees} showAttendees={showAttendees} attendees={attendees} updateEventAttendance={updateEventAttendance}/>
-            ))}
-          </div>
-        ) : null}
-      </>
+      <Attendees updateEventAttendance={updateEventAttendance} toggleEventAttendees={toggleEventAttendees} showAttendees={showAttendees} attendees={attendees} />
     </li>
   </>
   )
